==> step3_merge_all_responses.ipynb <==
# Step 3: Merge all prompts from all runs into one long-format dataset

Depends on: 

- 8 CSV files generated in  `processed_data/umg_parsed_queries_v2*.csv`

Outputs: 
- 1 large (395200 row) CSV file containing all results and metadata: `processed_data/umg_all_parsed_queries.csv`

import os
import json
from tqdm import tqdm
import pandas as pd
import glob
import datetime
pd.set_option('display.max_columns', None)

start = datetime.datetime.now()

models = [ "gpt-3.5-turbo-0613", "gpt-3.5-turbo-0125", "gpt-4-turbo-2024-04-09", "gpt-4o-2024-05-13"]

glob.glob("processed_data/umg_parsed_queries_v2*.csv")

df = pd.DataFrame()
for filen in glob.glob("processed_data/umg_parsed_queries_v2*.csv"):
    df_temp = pd.read_csv(filen)
    if len(df) == 0:
        df = df_temp
    else:
        df = pd.concat([df,df_temp])

majors_dict = {
    'Education': 'Social Sciences',
    'Psychology': 'Social Sciences',
    'Nursing': 'STEM',
    'Literature': 'Arts and Humanities',
    'History': 'Arts and Humanities',
    'Gender Studies': 'Arts and Humanities',
    'Electrical Engineering': 'STEM',
    'Data Science': 'STEM',
    'Biology': 'STEM',
    'Philosophy': 'Arts and Humanities',
    'Computer Science': 'STEM',
    'Economics': 'Social Sciences',
    'Communication': 'Social Sciences',
    'Neuroscience': 'STEM',
    'Business': 'Social Sciences',
    'Black Studies': 'Arts and Humanities',
    'Biomedical Sciences': 'STEM',
    'Visual Arts': 'Arts and Humanities',
    'Xyzzy': 'Fake-Control',
    'None-Control': 'None-Control'
}

# Add the 'major_type' column to the DataFrame
df['major_type'] = df['Major'].map(majors_dict)
df['major_type'].value_counts()

df.columns

df.sort_values(by=["model", "run_type","Pronoun","University","Major"], inplace=True)
df.reset_index(inplace=True)

df.to_csv("processed_data/umg_all_parsed_queries.csv", index=None)

df.describe()

end = datetime.datetime.now()
print("Elapsed time:", end-start)
