==> step4_0_performance_refusals.ipynb <==
# Step 4.0: Analysis: Refusal rate

Depends on: 
- CSV file generated in step 3 containing all results and metadata: `processed_data/umg_all_parsed_queries.csv`

Outputs: no files, only results in notebook

import os
import json
from tqdm import tqdm
import pandas as pd
import seaborn as sns
import glob
import matplotlib.pyplot as plt
import numpy as np
import scipy.stats as stats
from statsmodels.stats.multicomp import pairwise_tukeyhsd
import statsmodels.api as sm
import statsmodels.formula.api as smf
from statsmodels.formula.api import ols
pd.set_option('display.max_columns', None)
sns.set_style('whitegrid')

models = [ "gpt-3.5-turbo-0613", "gpt-3.5-turbo-0125", "gpt-4-turbo-2024-04-09", "gpt-4o-2024-05-13"]

df = pd.read_csv("processed_data/umg_all_parsed_queries.csv")
df

df['refusal'] = df['query_response_parsed'].apply(pd.isna)

df['content_len'] = df['content'].apply(len)

df.groupby(['model', 'run_type'])['refusal'].sum()

df.groupby(['model', 'run_type'])['refusal'].mean()*100

df.groupby(['model', 'run_type'])['content_len'].mean()

df['content_over_50'] = df['content_len'].apply(lambda x: x>50)
df['content_over_100'] = df['content_len'].apply(lambda x: x>100)

df.groupby(['model', 'run_type'])['content_len'].mean()

df.groupby(['model', 'run_type'])['content_over_100'].value_counts()

df.groupby(['model', 'run_type'])['content_over_100'].mean()*100

df[df['content_over_100']].sample(1).content.values

df[df['content_over_100'] & df['refusal']==True].sample(1).content.values
